apps/api/pyproject.toml
[build-system]
requires = ["setuptools>=69", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "atlas-api"
version = "0.1.0"
description = "Atlas backend API"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
  "alembic>=1.13.3",
  "duckdb>=1.1.3",
  "fastapi>=0.115.2",
  "numpy>=2.1.2",
  "optuna>=3.6.1",
  "pandas>=2.2.3",
  "pyarrow>=18.0.0",
  "pydantic-settings>=2.6.1",
  "python-multipart>=0.0.12",
  "redis>=5.2.0",
  "rq>=1.16.2",
  "sqlmodel>=0.0.22",
  "uvicorn[standard]>=0.32.0"
]

[project.optional-dependencies]
dev = [
  "httpx>=0.27.2",
  "mypy>=1.13.0",
  "pytest>=8.3.3",
  "pytest-cov>=5.0.0",
  "ruff>=0.7.1"
]

[tool.setuptools.packages.find]
where = ["."]
include = ["app*"]

[tool.ruff]
line-length = 100
target-version = "py312"

[tool.pytest.ini_options]
pythonpath = ["."]
testpaths = ["tests"]
